Note changes to CLI arguments
Are these fine here, or should we create a new section for CLI flags?

The `--dev`, `-d`, `--prod`, and `-p` flags technically belong to webpack-cli, so I'm not sure if they belong in this changelog. But still, people might run into this when migrating to webpack v5, since they'll update webpack-cli alongside it.

`--optimize-minimize` and `-p` are mentioned in this page: https://webpack.js.org/guides/production/#cli-alternatives so I assume they were active in v4. I'll update that doc too.